Adapted current code to wotk with Kloppy meta_data
Now that meta_data information is included in the dataset provided by
Kloppy, we implemented changes to the current models and example pattern
team_stretched.py to use this meta_data to get the team tracking data
frame without the goalkeeper.

The methods to get the dataframe of a team and get players_ids by team
were added to the DataPackage dataclass. This could change in the future,
and/or could be moved to Kloppy itself, but for now I think this is the
best place for it.