Setup testing of patterns, plus meta_data to metadata for Kloppy v1
Beside some mninor changes, the biggest change in this commit is setting
up the testing of patterns. There are still some things that need
improvement, but this will do for now.

How it works is that it load a pickle of a game_dataset object save
under `codeball\tests\files` and based on that runs the patterns in the
test file. This works fine, but: 1) this game_dataset object witghts
~300MB, so not really suited to be inlcuded in the repo and 2) more
game_datasets will be needed when we add events, or automated data etc.

Not a problem when I develop locally, but something we will need to
think about how to improve when we open source the repo and other people
need to test the code as well.