Divided common.py into game_dataset and patterns_set
There was no reason to have patterns information as an attribute of a
game_dataset class. It was prone to curcular imports. Plus, and more
importantly, this allows to separate the method to work on tactical
analysis of a game, from the methods to compute patterns.

Potentially they could even tbe different repositories, though for now
its better to keep it on one.

With this setup, all methods that compute tactical computations (like
get_set_pieces_for_team or find_stretched_intervals, etc) will be
methods of the GameDataset class, and Patterns and PatternAnalysis will
call them and work with them but should not have any private methods
that compute any tactical information.