Added passes into the box pattern and arrow viz
In this pattern we have added a prototype of detection of passes into
the box. Zone thresholds are implemented in the pattern. In time the
idea would be to have defined zones and properties so that events can be
quaery to the GameDatase. This event also pause at the moment of the
pass and shows an arrow to where the pass is going to end.