Refactor passes into the box + added tactical methods and models
In this commit I refectored the pattern *Passes into the box" by
creating tactical methods and models.

The models for now only contain
Zone, and Vertices, used to define relevant zones in the field. For
example, Zone.OPPONENT_BOX, and Vertices are the vertices that define
those zones. For now it only handles rectanagles.

The methos are methods associated to the Events (added toKloppy class
Event), and to the GameDataset class. To the Event I added methods
.is_pass() and .is_complete(). To the GameDataset, a method .pass() that
you can give a Zone and a Result as input. For now this methods only
handle the cases in set_pieces and passes_into_the_box, but will be
extended to handle all cases (inside, from, to, different results, etc).

A problem now is that the newly added methods for Event don't show in th
elinter. Not sure what's the best way to solve this.