Added new _enrich_data()_ method and attack direction to periods
There are several steps that need to be performed on the data after
loading to set it up for analysis. For example, computing the attack
directio on each half, transforming events to to left to right, compute
possessions, etc. For that step we have added a new `enrich_data()` method
to `game_dataset` class to run this steps.

For now, it only includes setting up the `attack_direction` (home-away
or away-home) for each half from the information gathered from the
tracking data.