Added inversion of events to the enrichment step
To be able to work better with events (for example when looking for
evnets from or to a certain area, etc) it's better to have all events
always oriented from left to right. This transformation which depends on
the attacking direction of each team on each period, is now added to the
data enrichment stage.

All events have been adapted to work with this, together with the api to
create an PatternEvent from an event. Before it was done by passing the
event.raw_event dict, not by passing the Event object.

There are certain improvements pending:
	1. To be able to better handle the difference between video and
	field coordiantes for inverted events.
	2. Now we only invert the coordinates of the event object, but
	not of the raw_event dict.
	3. All enrichment should be done before creating the dataframes
	from the data.