Refactor to Codeframe structures (TrackingDataFrames, EventDataFrames)
To be able to create an api to interact with the data that allows for
chainable methods, we needed to change our data structures from
DataPackages to Codeframes, which are subclasses of pandas DataFrames.
There are three codeframes, TrackingDataFrame, EventsDataFrame, and
PossessionsDataFrame.

This allows to use all common pandas dataframe methods, but also
codeball related methods that return indexes as pandas series or
filtered datasets.

Codeframes also have attributes metadata and data_type that persist when
filtering, and records to host the origianl events list that are only
available on the main/base codeframe.

There are some initial methids like .team(), .players(), .type() etc,
but still many to be devleped. All 3 existing patterns were adapted to
this new data structures and methods.

Also change the way to define a Zone, as a list of boxes.