Move data utils to game_dataset
There were two methods in utils/data_handlers. One to change from frame
to miliseconds, one to finde intervals. Since they were not attached to
the game_dataset class, so for example frame_rate had to be passed as an
argument.

I changed them to be methods of game_dataset now, so they can read
metadata directyly from game_dataset. One of the methods, (frame to
miliseconds) requiere for game_dataset to be available at the Pattern
base class. For some reason that was not the case, game_dataset was
passed only to the childs / subclasses when instantiating a pattern so I
changed that.