Make compatible with latest kloppy released version
I had implemented in my local kloppy repo a method to read the field
length as a property of the PitchDimensions model. That remains as a
beginner issue on Kloppy, so I rather leave that there. I changed the
code on the stretchd method to compute the length based on the available
attributes of the PitchDimensions class.