First full implementation of documentaiton ready
We are using mkdocs-material for the documentation of this repo. At the
moment we are not using any docstrings, but rather focus the efforts on
the markdown documentation.

We tried using a another package to read docstrings from the code, but
formating them to play along with the rest of the files proved
difficult.

There is also a CI workflow in place, to automatically push new changes
in the documentation when a new release is created on github. The
documentation is visible at codeball.metrica-sports.com.